recipes-tss/tss2/ibmtss2: Drop support for TPM1.2
On kirkstone compilation currently fails due to multiple definitions for
the same symbol in two different files. These files are Commands.c and
Commands12.c corresponding to TPM2.0 and TPM1.2 implementations
respectively.

A simple fix is to disable TPM1.2 support for this version of ibmtss2.
I would recommend bumping ibmtss2 to version 1.6.0 later in the kirkstone
migration process which should fix this problem itself and then drop
this version of ibmtss2. For now this is the lowest impact fix to get
compilation going again.

Signed-off-by: Johannes Wiesboeck <johannes.wiesboeck@aisec.fraunhofer.de>