recipes-connectivity/ssh: SSH keys for debug access to cml
In order to access the cml layer via SSH in development builds we need
to create keys for the server and for the client. For simplified access
we create these during the build process, install them to the image.
Generated keys are deployed to DEPLOY_DIR_IMAGE/ssh-keys. We set
SSTATE_SKIP_CREATION to 1 to avoid commiting any generated keys to
potential sstate mirrors.

We further require ssh-keygen as a installed host tool.

Signed-off-by: Johannes Wiesboeck <johannes.wiesboeck@aisec.fraunhofer.de>