images/trustx-cml-initramfs: Deploy kernel before build
Retrieving the kernel version requires the file
${STAGING_KERNEL_BUILDDIR}/kernel-abiversion to be exist. When the
kernel is checked out from sstate-cache this file is not created.
Forcing the kernel workdir to be deployed fixes this problem.

NB: Other recipes implement this mechanism by inheriting the kernelsrc
class which is similar but puts the dependency on the do_patch task,
that is not available in an image build. Therefore, we implement this by
hand, by manually depending on the do_shared_workdir from the do_rootfs
task.

Signed-off-by: Johannes Wiesboeck <johannes.wiesboeck@aisec.fraunhofer.de>